org search move back to original, fix image paths to be absolute
