+++ src/components/Site.tsx
@@ -8,9 +8,8 @@ import { Theme } from "../style/theme";
 import { CssBaseline, withStyles } from "@material-ui/core";
 import { TopNav } from "./nav/TopNav";
 import { SideNav } from "./nav/SideNav";
-import { SearchBar } from "./search/SearchBar";
-import { VolunteeringSearchBar } from "./volunteering/VolunteeringSearchBar";
 import { SearchBarMerged } from "./search/SearchBarMerged";
+import { SearchBar } from "./organizations/SearchBar";
 import { CelebrationModal } from "./modals/CelebrationModal";
 import { FacebookLoginLogout } from "./facebook/FacebookLoginLogout";
 import classNames from 'classnames';
@@ -29,7 +28,7 @@ export const SiteWithoutStyle = (props) => {
                     <Route exact path="/" component={HomePage} />
                     <Route path="/dev" component={() => <FacebookLoginLogout {...props} />} />
                     <Route path="/castleRisk" component={routerProps => <div><CastleRisk {...routerProps} {...props}/></div>} />
-                    <Route path="/organizations/search" render={() => <div style={{margin: "150px auto"}}><SearchBarMerged {...props} id='search' /></div>} />
+                    <Route path="/organizations/search" render={() => <div style={{margin: "150px auto"}}><SearchBar {...props} /></div>} />
                     <Route path="/volunteering/search" render={() => <div style={{margin: "150px auto"}}><SearchBarMerged {...props} id='volunteer' /></div>} />
                     <Route component={NotFound} />
                 </Switch>
